check for startup script

+++ static/fetch_lib.sh
@@ -27,8 +27,11 @@ then
         exit 1
     fi
 else
-    print_text_in_color "$ICyan" "Updating lib..."
-    curl -sfL https://raw.githubusercontent.com/nextcloud/vm/master/lib.sh -o /var/scripts/lib.sh
+    if ! [ -f /var/scripts/nextcloud-startup-script.sh ]
+    then
+        print_text_in_color "$ICyan" "Updating lib..."
+        curl -sfL https://raw.githubusercontent.com/nextcloud/vm/master/lib.sh -o /var/scripts/lib.sh
+    fi
 fi
 
 # shellcheck source=lib.sh
